backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-jose
